Arman/fix/improvment issues

See merge request atlas39/jahizan!43

+++ app/admin/eav/entityTypes/fields/[id]/values/page.tsx
@@ -2,7 +2,7 @@
 import { fetcher, useFetcher } from "@/app/components/global/fetcher";
 import Loading from "@/app/components/global/loading";
 import { DataGrid, GridColDef } from "@mui/x-data-grid";
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { useAtom } from "jotai";
 import { toast } from "react-toastify";
 import Image from "next/image";
@@ -11,10 +11,11 @@ import LightDataGrid from "@/app/components/global/LightDataGrid/LightDataGrid";
 import { Button, IconButton } from "@mui/material";
 import ModeEditIcon from "@mui/icons-material/ModeEdit";
 import DeleteIcon from "@mui/icons-material/Delete";
+import Swal from "sweetalert2";
 
 export default function Eav({ params }) {
   const [title, setTitle] = useAtom(pageTitle);
-
+  const [triggered, setTriggered] = useState(false);
   useEffect(() => {
     setTitle({
       title: "مقادیر",
@@ -25,12 +26,25 @@ export default function Eav({ params }) {
 
   const deleteValue = async (id) => {
     try {
-      const req = await fetcher({
-        url: `/v1/api/eav/admin/attributeValues/${id}`,
-        method: "DELETE",
+      const result = await Swal.fire({
+        title: "مطمئن هستید؟",
+        text: "با حذف این گزینه امکان بازگشت آن وجود ندارد",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "بله حذفش کن",
+        cancelButtonText: "لغو",
       });
-      toast.success("موفق");
-      categoriesRefetch();
+
+      if (result.isConfirmed) {
+        const req = await fetcher({
+          url: `/v1/api/eav/admin/attributeValues/${id}`,
+          method: "DELETE",
+        });
+        toast.success("موفق");
+        setTriggered(!triggered);
+      }
     } catch (error) {
       toast.error(error.message);
     }
@@ -55,12 +69,7 @@ export default function Eav({ params }) {
       accessorKey: "Actions",
       header: "عملیات",
       size: 200,
-      muiTableHeadCellProps: {
-        align: "right",
-      },
-      muiTableBodyCellProps: {
-        align: "right",
-      },
+
       Cell: ({ row }) => (
         <>
           <a
@@ -85,6 +94,7 @@ export default function Eav({ params }) {
       <LightDataGrid
         url={`/v1/api/eav/admin/attributeValues?sortOrder=ASC&orderBy=id&ignorePaging=false&attributeId=${params.id}`}
         columns={columns}
+        triggered={triggered}
       />
     </div>
   );
