+++ app/admin/ecommerce/discounts/conditions/[id]/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { DataGrid, GridColDef } from "@mui/x-data-grid";
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { useAtom } from "jotai";
 import { toast } from "react-toastify";
 import { fetcher, useFetcher } from "@/app/components/global/fetcher";
@@ -11,9 +11,10 @@ import { IconButton } from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
 import ModeEditIcon from "@mui/icons-material/ModeEdit";
 import SaveBar from "@/app/components/global/SaveBar";
-
+import Swal from "sweetalert2";
 export default function DiscountConditions({ params }) {
   const [title, setTitle] = useAtom(pageTitle);
+  const [triggered, setTriggered] = useState(false);
 
   useEffect(() => {
     setTitle({
@@ -25,12 +26,25 @@ export default function DiscountConditions({ params }) {
 
   const deleteRow = async (id) => {
     try {
-      const req = await fetcher({
-        url: `/v1/api/ecommerce/admin/discountConditions/${id}`,
-        method: "DELETE",
+      const result = await Swal.fire({
+        title: "مطمئن هستید؟",
+        text: "با حذف این گزینه امکان بازگشت آن وجود ندارد",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "بله حذفش کن",
+        cancelButtonText: "لغو",
       });
-      toast.success("موفق");
-      refetch();
+
+      if (result.isConfirmed) {
+        const req = await fetcher({
+          url: `/v1/api/ecommerce/admin/discountConditions/${id}`,
+          method: "DELETE",
+        });
+        toast.success("موفق");
+        setTriggered(!triggered);
+      }
     } catch (error) {
       toast.error(error.message);
     }
@@ -45,53 +59,6 @@ export default function DiscountConditions({ params }) {
     `/v1/api/ecommerce/admin/discountConditions?sortOrder=DESC&discountId=${params.id}&offset=0&limit=10&orderBy=id`,
     "GET"
   );
-
-  // const columns: GridColDef[] = [
-  //   {
-  //     field: "id",
-  //     headerName: "شناسه",
-  //     width: 150,
-  //   },
-  //   {
-  //     field: "name",
-  //     headerName: "نام ",
-  //     width: 150,
-  //   },
-  //   {
-  //     field: "hexCode",
-  //     headerName: "کد رنگ ",
-  //     width: 150,
-  //     renderCell: ({ row }) => (
-  //       <div className="flex items-center">
-  //         <span
-  //           style={{ background: row.hexCode }}
-  //           className={`w-5 h-5 ml-2 rounded-sm`}
-  //         ></span>
-  //         {row.hexCode}
-  //       </div>
-  //     ),
-  //   },
-  //   {
-  //     field: "list",
-  //     headerName: "ویرایش",
-  //     width: 450,
-  //     renderCell: (row) => (
-  //       <>
-  //         <a onClick={(e) => deleteItem(row.id)}>
-  //           <button
-  //             type="button"
-  //             className="focus:outline-none mx-4 text-white bg-red-600 hover:bg-red-800 focus:ring-4 focus:ring-purple-300 font-medium rounded-lg text-sm px-5 py-2.5 dark:bg-purple-600 dark:hover:bg-purple-700 dark:focus:ring-purple-900"
-  //           >
-  //             حذف
-  //           </button>
-  //         </a>
-  //       </>
-  //     ),
-  //   },
-  // ];
-  // if (conditionsIsLoading) {
-  //   return <Loading />;
-  // }
 
   const columns = [
     {
@@ -121,37 +88,14 @@ export default function DiscountConditions({ params }) {
       maxSize: 400, //max size enforced during resizing
       size: 400, //medium column
     },
-    // {
-    //   accessorKey: "hexCode",
-    //   header: "کد رنگ ",
-    //   size: 20,
-    //   Cell: ({ row }) => (
-    //     <div className="flex items-center">
-    //       <span
-    //         style={{ background: row.original.hexCode }}
-    //         className={`w-5 h-5 ml-2 rounded-sm`}
-    //       ></span>
-    //       {row.original.hexCode}
-    //     </div>
-    //   ),
-    // },
+
     {
       accessorKey: "Actions",
       header: "عملیات",
       size: 200,
-      muiTableHeadCellProps: {
-        align: "right",
-      },
-      muiTableBodyCellProps: {
-        align: "right",
-      },
+
       Cell: ({ row }) => (
         <>
-          {/* <a href={`/admin/ecommerce/discounts/${row.id}`}>
-            <IconButton aria-label="delete" color="primary">
-              <ModeEditIcon />
-            </IconButton>
-          </a> */}
           <a onClick={(e) => deleteRow(row.id)}>
             <IconButton aria-label="delete" color="error">
               <DeleteIcon />
@@ -166,6 +110,7 @@ export default function DiscountConditions({ params }) {
       <LightDataGrid
         url={`/v1/api/ecommerce/admin/discountConditions?sortOrder=DESC&discountId=${+params.id}&offset=0&limit=10&orderBy=id`}
         columns={columns}
+        triggered={triggered}
       />
     </div>
   );
